requirements.txt
Django==1.5.4
## needed for ImageField
PIL==1.1.7
## spam filter
akismet==0.2.0

argparse==1.2.1
distribute==0.7.3

## comments
django-crispy-forms==1.4.0
django-fluent-comments==0.9.2

## tagging support
django-taggit==0.10

## FIXME: could not find svn URL in dependency_links for this package:
=======
flup==1.0.3.dev-20110405
sorl-thumbnail==11.12
wsgiref==0.1.2
